backward compat rotation and stuff

+++ fabric/src/main/java/net/mehvahdjukaar/supplementaries/integration/fabric/CCCompatImpl.java
@@ -5,7 +5,7 @@ import dan200.computercraft.api.peripheral.IPeripheral;
 import dan200.computercraft.api.peripheral.PeripheralLookup;
 import dan200.computercraft.shared.media.items.PrintoutData;
 import dan200.computercraft.shared.media.items.PrintoutItem;
-import net.mehvahdjukaar.supplementaries.common.block.cannon.EntityAngles;
+import net.mehvahdjukaar.moonlight.api.util.math.EntityAngles;
 import net.mehvahdjukaar.supplementaries.common.block.tiles.CannonBlockTile;
 import net.mehvahdjukaar.supplementaries.common.block.tiles.SpeakerBlockTile;
 import net.mehvahdjukaar.supplementaries.reg.ModRegistry;
@@ -153,7 +153,7 @@ public class CCCompatImpl {
         public void setYaw(double value) {
             Quaternionf orientation = tile.getLocalOrientation(1);
             EntityAngles angles = EntityAngles.fromQuaternion(orientation);
-            angles = angles.withYaw(value);
+            angles = angles.withYaw((float) value);
 
             tile.setLocalOrientation(angles.toQuaternion());
             tile.syncToClients(false);
@@ -163,7 +163,7 @@ public class CCCompatImpl {
         public void setPitch(double value) {
             Quaternionf orientation = tile.getLocalOrientation(1);
             EntityAngles angles = EntityAngles.fromQuaternion(orientation);
-            angles = angles.withPitch(value);
+            angles = angles.withPitch((float) value);
 
             tile.setLocalOrientation(angles.toQuaternion());
